backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP and web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Graph and data processing
networkx==3.2.1
pandas==2.1.4
numpy==1.26.3

# AI and NLP
google-genai
scikit-learn==1.4.0
nltk==3.8.1

# Reddit API
praw==7.7.1

# Utilities
python-dotenv==1.0.0
